Added JWT Authorization

+++ backEnd/src/services/userService.ts
@@ -6,8 +6,13 @@ const logger = require('../loggers/logger');
 
 export const getUserService = async (req: any, res: Response) => {
   try {
-    // const result = await User.find();
-    // res.json({ data: result });
+    let requestedUser = await User.findById(req.decoded.id)
+    if (!requestedUser) {
+      return res.status(401).send("Cannot find the user")
+    }
+    if (requestedUser.type != "Admin") {
+      return res.status(401).send("Only Admins can request.")
+    }
 
     const page: any = req.query.page ? req.query.page - 1 : 0;
     const usersPerPage: any = req.query.chunk || 5;
@@ -19,7 +24,6 @@ export const getUserService = async (req: any, res: Response) => {
     });
 
   } catch (err) {
-    console.log(err)
     logger.userErrorLogger.log('info', 'Error User Lists')
     return res.status(400).send("An Error occured in get user");
   }
@@ -27,6 +31,13 @@ export const getUserService = async (req: any, res: Response) => {
 
 export const createUserService = async (req: any, res: Response) => {
   try {
+    let requestedUser = await User.findById(req.decoded.id)
+    if (!requestedUser) {
+      return res.status(401).send("Cannot find the user")
+    }
+    if (requestedUser.type != "Admin") {
+      return res.status(401).send("Only Admins can request.")
+    }
     let profile = req.body.profileImage;
     if (req.file) {
       profile = req.file.path.replace('\\', '/');
@@ -44,31 +55,42 @@ export const createUserService = async (req: any, res: Response) => {
     }
     const post = new User(userCreate);
     const result = await post.save();
-    res.status(200).json({ msg: "Created User Successfully!!", data: result, status: 1 });
+    return res.status(200).json({ msg: "Created User Successfully!!", data: result, status: 1 });
   } catch (err) {
-    res.send("An Error occured in create user");
-    console.log(err)
     logger.userInfoLogger.log('info', 'Error Create User')
+    return res.status(400).send("An Error occured in create user");
   }
 };
 
 export const findUserService = async (req: any, res: Response) => {
   try {
-    const findData = await User.findById(req.params.id)
-    res.send({ data: findData })
+    let requestedUser = await User.findById(req.decoded.id)
+    if (!requestedUser) {
+      return res.status(401).send("Cannot find the user")
+    }
+    const findData: any = await User.findById(req.params.id)
+    if (requestedUser._id != findData._id && requestedUser.type != "Admin") {
+      return res.status(403).send("Not Authorized")
+    }
+    return res.status(200).send({ data: findData })
   } catch (err) {
-    console.log(err)
-    res.send("An Error occured in find user");
     logger.userErrorLogger.log('info', 'Error User Not Found')
+    return res.send("An Error occured in find user");
   }
 };
 
 export const updateUserService = async (req: any, res: Response) => {
   try {
+    let requestedUser = await User.findById(req.decoded.id)
+    if (!requestedUser) {
+      return res.status(401).send("Cannot find the user")
+    }
     const user = await User.findById(req.params.id);
     if (!user) {
-      const error = new Error("Not Found");
-      throw error;
+      return res.status(401).send("Cannot find the user")
+    }
+    if (requestedUser._id != user._id && requestedUser.type != "Admin") {
+      return res.status(403).send("Not Authorized")
     }
     let profile = req.body.profileImage;
     if (req.file) {
@@ -88,8 +110,8 @@ export const updateUserService = async (req: any, res: Response) => {
     user.created_user_id = req.body.created_user_id;
     user.updated_user_id = req.body.updated_user_id;
     const result = await user.save();
-    res.json({
-      msg: "Image Updated Successfully", data: {
+    return res.status(200).json({
+      msg: "User Updated Successfully", data: {
         _id: result._id,
         profile: result.profile,
         name: result.name,
@@ -101,32 +123,32 @@ export const updateUserService = async (req: any, res: Response) => {
       }
     });
   } catch (err) {
-    console.log(err)
-    res.send("an error occured in Edit User");
     logger.userErrorLogger.log('info', 'Error Update User')
+    return res.status(400).send("an error occured in Edit User");
   }
 };
 
 export const deleteUserService = async (req: any, res: Response) => {
   try {
+    let requestedUser = await User.findById(req.decoded.id)
+    if (!requestedUser) {
+      return res.status(401).send("Cannot find the user")
+    }
     const user: any = await User.findById(req.params.id);
 
     if (!user) {
-      const error: any = new Error("Not Found !!")
-      error.statusCode = 404;
-      throw error;
+      return res.status(401).send("Cannot find the user")
+    }
+    if (requestedUser._id != user._id && requestedUser.type != "Admin") {
+      return res.status(403).send("Not Authorized")
     }
 
-    const result = await user.save();
-    user.deleted_at = new Date();
     await User.findByIdAndDelete(req.params.id);
 
-    res.json({ msg: "Deleted User Successfully", data: result })
-    // res.json({ message: "User with id " + req.params._id + " removed." })
+    return res.status(200).send("User Deleted Successfully")
 
   } catch (err) {
-    res.send("An Error Occured During Delete user")
-    console.log(err)
     logger.userErrorLogger.log('info', 'Error Delete User')
+    return res.status(400).send("An Error Occured During Delete user")
   }
 };
